Refactored vocab list query to use projection extension methods.
