+++ GermanVocabApp.DataAccess.EntityFramework/Projection/VocabListItemProjectionExtensions.cs
@@ -1,19 +1,33 @@
 ﻿using GermanVocabApp.DataAccess.EntityFramework.Models;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
 
 namespace GermanVocabApp.DataAccess.EntityFramework.Projection;
 
-public static class VocabListProjectionExtensions
+public static class VocabListItemProjectionExtensions
 {
-    public static IQueryable<VocabListInfoDto> ProjectToInfoDto(this IQueryable<VocabList> query)
+    public static IEnumerable<VocabListItem> ProjectToFullItem(this IEnumerable<VocabListItem> query)
     {
-        IQueryable<VocabListInfoDto> infoQuery;
-        infoQuery = query.Select(vl => new VocabListInfoDto()
+        return query.Select(li => new VocabListItem()
         {
-            Id = vl.Id,
-            Name = vl.Name,
-            Description = vl.Description,
+            Id = li.Id,
+            WordType = li.WordType,
+            IsWeakMasculineNoun = li.IsWeakMasculineNoun,
+            ReflexiveCase = li.ReflexiveCase,
+            Separability = li.Separability,
+            Transitivity = li.Transitivity,
+            ThirdPersonPresent = li.ThirdPersonPresent,
+            ThirdPersonImperfect = li.ThirdPersonImperfect,
+            AuxiliaryVerb = li.AuxiliaryVerb,
+            Perfect = li.Perfect,
+            Gender = li.Gender,
+            German = li.German,
+            Plural = li.Plural,
+            Preposition = li.Preposition,
+            PrepositionCase = li.PrepositionCase,
+            Comparative = li.Comparative,
+            Superlative = li.Superlative,
+            English = li.English,
+            VocabListId = li.VocabListId,
+            FixedPlurality = li.FixedPlurality,
         });
-        return infoQuery;
     }
 }
